refactor: simplify error handling with constants and improve error messages
## CHANGES

- Replace hardcoded error codes with MCP constants
- Add error code parameter to raise_mcp_error utility
- Improve error messages with operation context
- Update raise_mcp_error to return NoReturn type
- Consolidate duplicate error handling logic
- Use INTERNAL_ERROR and INVALID_PARAMS constants
- Make error messages more descriptive and consistent
