+++ src/fabric_mcp/core.py
@@ -1,51 +1,30 @@
 """Core MCP server implementation using the Model Context Protocol."""
 
-import fnmatch
-import json
 import logging
 from asyncio.exceptions import CancelledError
 from collections.abc import Generator
-from dataclasses import dataclass
-from typing import Any, cast
+from typing import Any
 
 import httpx
 from anyio import WouldBlock
 from fastmcp import FastMCP
-from mcp.shared.exceptions import McpError
-from mcp.types import ErrorData
+from mcp.types import INTERNAL_ERROR, INVALID_PARAMS
 
 from . import __version__
-from .api_client import FabricApiClient
+from .api_client import FabricApiClient  # Re-export for test compatibility
 from .config import get_default_model
+from .constants import DEFAULT_MCP_HTTP_PATH, DEFAULT_MODEL, DEFAULT_VENDOR
+from .fabric_tools import FabricToolsMixin
+from .models import PatternExecutionConfig
+from .sse_parser import SSEParserMixin
+from .utils import raise_mcp_error
+from .validation import ValidationMixin
 
-DEFAULT_MCP_HTTP_PATH = "/message"
-
-DEFAULT_VENDOR = "openai"
-DEFAULT_MODEL = "gpt-4o"  # Default model if none specified in config
-
-# Sensitive configuration key patterns for redaction
-SENSITIVE_CONFIG_PATTERNS = ["*_API_KEY", "*_TOKEN", "*_SECRET", "*_PASSWORD"]
-
-# API key prefixes that indicate a value is an API key regardless of key name
-API_KEY_PREFIXES = ["sk-", "ant-", "xai-", "gsk_", "AIza"]
+# Re-export for test compatibility
+__all__ = ["FabricMCP", "FabricApiClient"]
 
 
-@dataclass
-class PatternExecutionConfig:  # pylint: disable=too-many-instance-attributes
-    """Configuration for pattern execution parameters."""
-
-    model_name: str | None = None
-    vendor_name: str | None = None
-    strategy_name: str | None = None
-    variables: dict[str, str] | None = None
-    attachments: list[str] | None = None
-    temperature: float | None = None
-    top_p: float | None = None
-    presence_penalty: float | None = None
-    frequency_penalty: float | None = None
-
-
-class FabricMCP(FastMCP[None]):
+class FabricMCP(FastMCP[None], FabricToolsMixin, SSEParserMixin, ValidationMixin):
     """Base class for the Model Context Protocol server."""
 
     def __init__(self, log_level: str = "INFO"):
@@ -105,146 +84,114 @@ class FabricMCP(FastMCP[None]):
                 e,
             )
 
-    def _make_fabric_api_request(
+    def get_vendor_and_model(self, config: PatternExecutionConfig) -> tuple[str, str]:
+        """Get the vendor and model based on the provided configuration."""
+        vendor_name = config.vendor_name or self._default_vendor
+        if not vendor_name:
+            logger = logging.getLogger(__name__)
+            logger.debug(
+                "Vendor name is None or empty. Set to hardcoded default vendor: %s",
+                DEFAULT_VENDOR,
+            )
+            vendor_name = DEFAULT_VENDOR
+
+        model_name = config.model_name or self._default_model
+        if not model_name:
+            logger = logging.getLogger(__name__)
+            logger.debug(
+                "Model name is None or empty. Set to hardcoded default model: %s",
+                DEFAULT_MODEL,
+            )
+            model_name = DEFAULT_MODEL
+
+        return vendor_name, model_name
+
+    def _execute_fabric_pattern(
         self,
-        endpoint: str,
-        pattern_name: str | None = None,
-        operation: str = "API request",
-    ) -> Any:
-        """Make a request to the Fabric API with consistent error handling.
-
-        Args:
-            endpoint: The API endpoint to call (e.g., "/patterns/names")
-            pattern_name: Pattern name for pattern-specific error messages
-            operation: Description of the operation for error messages
-
-        Returns:
-            The parsed JSON response from the API
-
-        Raises:
-            McpError: For any API errors, connection issues, or parsing problems
+        pattern_name: str,
+        input_text: str,
+        config: PatternExecutionConfig | None,
+        stream: bool = False,
+    ) -> dict[Any, Any] | Generator[dict[str, Any], None, None]:
         """
-        try:
-            api_client = FabricApiClient()
-            try:
-                response = api_client.get(endpoint)
-                return response.json()
-            finally:
-                api_client.close()
-        except httpx.RequestError as e:
-            raise McpError(
-                ErrorData(
-                    code=-32603,  # Internal error
-                    message="Failed to connect to Fabric API",
-                )
-            ) from e
-        except httpx.HTTPStatusError as e:
-            if e.response.status_code == 500 and pattern_name:
-                # Check for pattern not found (500 with file not found message)
-                error_message = e.response.text or ""
-                if "no such file or directory" in error_message:
-                    raise McpError(
-                        ErrorData(
-                            code=-32602,  # Invalid params - pattern doesn't exist
-                            message=f"Pattern '{pattern_name}' not found",
-                        )
-                    ) from e
-                # Other 500 errors for pattern requests
-                raise McpError(
-                    ErrorData(
-                        code=-32603,  # Internal error
-                        message=f"Fabric API internal error: {error_message}",
-                    )
-                ) from e
-            # Generic HTTP status errors
-            status_code = e.response.status_code
-            reason = e.response.reason_phrase or "Unknown error"
-            raise McpError(
-                ErrorData(
-                    code=-32603,  # Internal error
-                    message=f"Fabric API error: {status_code} {reason}",
-                )
-            ) from e
-        except Exception as e:
-            raise McpError(
-                ErrorData(
-                    code=-32603,  # Internal error
-                    message=f"Unexpected error during {operation}: {e}",
-                )
-            ) from e
+        Execute a Fabric pattern against the API.
 
-    def fabric_list_patterns(self) -> list[str]:
-        """Return a list of available fabric patterns."""
-        response_data = self._make_fabric_api_request(
-            "/patterns/names", operation="retrieving patterns"
-        )
+        Separated from the tool method to reduce complexity.
+        """
+        # AC5: Client-side validation
+        if not pattern_name or not pattern_name.strip():
+            raise ValueError("pattern_name is required and cannot be empty")
 
-        # Validate response data type
-        if not isinstance(response_data, list):
-            raise McpError(
-                ErrorData(
-                    code=-32603,  # Internal error
-                    message="Invalid response from Fabric API: "
-                    "expected list of patterns",
-                )
-            )
+        # Use default config if none provided
+        if config is None:
+            config = PatternExecutionConfig()
 
-        # Cast to expected type
-        response_data = cast(list[Any], response_data)
+        vendor, model_name = self.get_vendor_and_model(config)
 
-        for item in response_data:
-            # Ensure each item is a string
-            if not isinstance(item, str):
-                raise McpError(
-                    ErrorData(
-                        code=-32603,  # Internal error
-                        message="Invalid pattern name in response: "
-                        f"expected string, got {type(item).__name__}",
-                    )
-                )
-
-        patterns = cast(list[str], response_data)
-
-        return patterns
-
-    def fabric_get_pattern_details(self, pattern_name: str) -> dict[str, str]:
-        """Retrieve detailed information for a specific Fabric pattern."""
-        # Use helper method for API request with pattern-specific error handling
-        response_data = self._make_fabric_api_request(
-            f"/patterns/{pattern_name}",
-            pattern_name=pattern_name,
-            operation="retrieving pattern details",
-        )
-
-        # Validate response data type
-        if not isinstance(response_data, dict):
-            raise McpError(
-                ErrorData(
-                    code=-32603,  # Internal error
-                    message="Invalid response from Fabric API: "
-                    "expected dict for pattern details",
-                )
-            )
-
-        response_data = cast(dict[str, Any], response_data)
-
-        # Validate required fields in the response
-        if not all(key in response_data for key in ("Name", "Description", "Pattern")):
-            raise McpError(
-                ErrorData(
-                    code=-32603,  # Internal error
-                    message="Invalid pattern details response: missing required fields",
-                )
-            )
-
-        # Transform Fabric API response to MCP expected format
-        details = {
-            "name": response_data["Name"],
-            "description": response_data["Description"],
-            "system_prompt": response_data["Pattern"],
+        # AC3: Construct proper JSON payload for Fabric API /chat endpoint
+        prompt_data: dict[str, Any] = {
+            "userInput": input_text,
+            "patternName": pattern_name.strip(),
+            "model": model_name,
+            "vendor": vendor,
+            "contextName": "",
+            "strategyName": config.strategy_name or "",
         }
 
-        return details
+        # Add variables if provided
+        if config.variables is not None:
+            prompt_data["variables"] = config.variables
+
+        # Add attachments if provided
+        if config.attachments is not None:
+            prompt_data["attachments"] = config.attachments
+
+        request_payload = {
+            "prompts": [prompt_data],
+            "language": "en",
+            "temperature": config.temperature
+            if config.temperature is not None
+            else 0.7,
+            "topP": config.top_p if config.top_p is not None else 0.9,
+            "frequencyPenalty": config.frequency_penalty
+            if config.frequency_penalty is not None
+            else 0.0,
+            "presencePenalty": config.presence_penalty
+            if config.presence_penalty is not None
+            else 0.0,
+        }
+
+        # AC1: Use FabricApiClient to call Fabric's /chat endpoint
+        api_client = FabricApiClient()
+        try:
+            # AC4: Handle Server-Sent Events (SSE) stream response
+            response = api_client.post("/chat", json_data=request_payload)
+            response.raise_for_status()  # Raise HTTPError for bad responses
+
+            if stream:
+                # Return generator for streaming mode
+                return self._parse_sse_stream(response)
+            # Return accumulated result for non-streaming mode
+            return self._parse_sse_response(response)
+
+        except httpx.ConnectError as e:
+            logger = logging.getLogger(__name__)
+            logger.error("Failed to connect to Fabric API: %s", e)
+            raise ConnectionError(f"Unable to connect to Fabric API: {e}") from e
+        except httpx.HTTPStatusError as e:
+            logger = logging.getLogger(__name__)
+            logger.error("Fabric API HTTP error: %s", e)
+            error_text = e.response.text
+            status_code = e.response.status_code
+            raise RuntimeError(
+                f"Fabric API returned error {status_code}: {error_text}"
+            ) from e
+        except Exception as e:
+            logger = logging.getLogger(__name__)
+            logger.error("Unexpected error calling Fabric API: %s", e)
+            raise RuntimeError(f"Unexpected error executing pattern: {e}") from e
+        finally:
+            api_client.close()
 
     def fabric_run_pattern(
         self,
@@ -333,232 +280,45 @@ class FabricMCP(FastMCP[None]):
                 "Fabric API returned error 500" in error_message
                 and "no such file or directory" in error_message
             ):
-                raise McpError(
-                    ErrorData(
-                        code=-32602,  # Invalid params - pattern doesn't exist
-                        message=f"Pattern '{pattern_name}' not found",
-                    )
-                ) from e
+                raise_mcp_error(
+                    e, INVALID_PARAMS, f"Pattern '{pattern_name}' not found"
+                )
             # Check for other HTTP status errors
             if "Fabric API returned error" in error_message:
-                raise McpError(
-                    ErrorData(
-                        code=-32603,  # Internal error
-                        message=f"Error executing pattern '{pattern_name}': {e}",
-                    )
-                ) from e
+                raise_mcp_error(
+                    e,
+                    INTERNAL_ERROR,
+                    f"Error executing pattern '{pattern_name}': {error_message}",
+                )
             # Other runtime errors
-            raise McpError(
-                ErrorData(
-                    code=-32603,  # Internal error
-                    message=f"Error executing pattern '{pattern_name}': {e}",
-                )
-            ) from e
+            raise_mcp_error(
+                e,
+                INTERNAL_ERROR,
+                f"Error executing pattern '{pattern_name}': {error_message}",
+            )
         except ConnectionError as e:
-            raise McpError(
-                ErrorData(
-                    code=-32603,  # Internal error
-                    message=f"Error executing pattern '{pattern_name}': {e}",
-                )
-            ) from e
+            raise_mcp_error(
+                e,
+                INTERNAL_ERROR,
+                f"Error executing pattern '{pattern_name}': {e}",
+            )
         except ValueError as e:
-            raise McpError(
-                ErrorData(
-                    code=-32602,  # Invalid params
-                    message=f"Invalid parameter for pattern '{pattern_name}': {e}",
-                )
-            ) from e
-
-    def fabric_list_models(self) -> dict[Any, Any]:
-        """Retrieve configured Fabric models by vendor."""
-        response_data = self._make_fabric_api_request(
-            "/models/names", operation="retrieving models"
-        )
-
-        # Validate response data type
-        if not isinstance(response_data, dict):
-            raise McpError(
-                ErrorData(
-                    code=-32603,  # Internal error
-                    message=(
-                        "Invalid response from Fabric API: expected dict for models"
-                    ),
-                )
+            raise_mcp_error(
+                e,
+                INVALID_PARAMS,
+                f"Invalid parameter for pattern '{pattern_name}': {e}",
             )
 
-        response_data = cast(dict[str, Any], response_data)
-
-        # Validate models field
-        models = response_data.get("models", [])
-        if not isinstance(models, list):
-            raise McpError(
-                ErrorData(
-                    code=-32603,  # Internal error
-                    message="Invalid models field: expected list",
-                )
-            )
-
-        # Validate each model name is a string
-        models_list = cast(list[Any], models)
-        for item in models_list:
-            if not isinstance(item, str):
-                raise McpError(
-                    ErrorData(
-                        code=-32603,  # Internal error
-                        message=(
-                            "Invalid model name in response: "
-                            f"expected string, got {type(item).__name__}"
-                        ),
-                    )
-                )
-
-        # Validate vendors field
-        vendors = response_data.get("vendors", {})
-        if not isinstance(vendors, dict):
-            raise McpError(
-                ErrorData(
-                    code=-32603,  # Internal error
-                    message="Invalid vendors field: expected dict",
-                )
-            )
-
-        # Validate vendor structure - each vendor should have a list of strings
-        vendors_dict = cast(dict[Any, Any], vendors)
-        for vendor_name, vendor_models in vendors_dict.items():
-            if not isinstance(vendor_name, str):
-                raise McpError(
-                    ErrorData(
-                        code=-32603,  # Internal error
-                        message=(
-                            "Invalid vendor name in response: "
-                            f"expected string, got {type(vendor_name).__name__}"
-                        ),
-                    )
-                )
-            if not isinstance(vendor_models, list):
-                raise McpError(
-                    ErrorData(
-                        code=-32603,  # Internal error
-                        message=(
-                            f"Invalid models list for vendor '{vendor_name}': "
-                            "expected list"
-                        ),
-                    )
-                )
-            vendor_models_list = cast(list[Any], vendor_models)
-            for model in vendor_models_list:
-                if not isinstance(model, str):
-                    raise McpError(
-                        ErrorData(
-                            code=-32603,  # Internal error
-                            message=(
-                                f"Invalid model name for vendor '{vendor_name}': "
-                                f"expected string, got {type(model).__name__}"
-                            ),
-                        )
-                    )
-
-        # Return validated structure
-        return {
-            "models": cast(list[str], models),
-            "vendors": cast(dict[str, list[str]], vendors),
-        }
-
-    def fabric_list_strategies(self) -> dict[Any, Any]:
-        """Retrieve available Fabric strategies."""
-        # Use helper method for API request
-        response_data = self._make_fabric_api_request(
-            "/strategies", operation="retrieving strategies"
-        )
-
-        # Validate response data type
-        if not isinstance(response_data, list):
-            raise McpError(
-                ErrorData(
-                    code=-32603,  # Internal error
-                    message="Invalid response from Fabric API: "
-                    "expected list of strategies",
-                )
-            )
-
-        response_data = cast(list[Any], response_data)
-        # Ensure all items are dictionaries
-        for item in response_data:
-            if not isinstance(item, dict):
-                raise McpError(
-                    ErrorData(
-                        code=-32603,  # Internal error
-                        message="Invalid strategy object in response: "
-                        f"expected dict, got {type(item).__name__}",
-                    )
-                )
-
-        # Cast to expected type
-        response_data = cast(list[dict[str, Any]], response_data)
-
-        # Validate each strategy object and build response
-        validated_strategies: list[dict[str, str]] = []
-        for item in response_data:
-            name = item.get("name", "")
-            description = item.get("description", "")
-            prompt = item.get("prompt", "")
-
-            # Type check all fields as strings and ensure name/description not empty
-            if (
-                isinstance(name, str)
-                and isinstance(description, str)
-                and isinstance(prompt, str)
-                and name.strip()  # Name must not be empty/whitespace
-                and description.strip()  # Description must not be empty/whitespace
-                # Note: prompt can be empty string - that's valid
-            ):
-                validated_strategies.append(
-                    {"name": name, "description": description, "prompt": prompt}
-                )
-            else:
-                # Log warning but continue with valid strategies
-                self.logger.warning(
-                    "Strategy object missing required string fields: %s",
-                    cast(Any, item),
-                )
-
-        return {"strategies": validated_strategies}
-
-    def fabric_get_configuration(self) -> dict[Any, Any]:
-        """Retrieve Fabric configuration with sensitive values redacted.
+    def get_default_model_config(self) -> tuple[str | None, str | None]:
+        """Get the current default model configuration.
 
         Returns:
-            dict[Any, Any]: Configuration key-value pairs with sensitive values
-            redacted. Sensitive keys (matching *_API_KEY, *_TOKEN, *_SECRET,
-            *_PASSWORD patterns) with non-empty values are replaced with
-            "[REDACTED_BY_MCP_SERVER]". Empty sensitive values and all
-            non-sensitive values are passed through unchanged.
+            Tuple of (default_model, default_vendor). Either or both can be None.
 
-        Raises:
-            McpError: For any API errors, connection issues, or invalid responses.
+        Note:
+            This method is primarily intended for testing and introspection.
         """
-        # Get configuration from Fabric API
-        response_data = self._make_fabric_api_request(
-            "/config", operation="retrieving configuration"
-        )
-
-        # Validate response data type
-        if not isinstance(response_data, dict):
-            raise McpError(
-                ErrorData(
-                    code=-32603,  # Internal error
-                    message=(
-                        "Invalid response from Fabric API: expected dict for config"
-                    ),
-                )
-            )
-
-        response_data = cast(dict[str, Any], response_data)
-
-        # Apply redaction to sensitive values
-        redacted_config = self._redact_sensitive_config_values(response_data)
-
-        return redacted_config
+        return self._default_model, self._default_vendor
 
     def http_streamable(
         self,
@@ -581,507 +341,3 @@ class FabricMCP(FastMCP[None]):
         except (KeyboardInterrupt, CancelledError, WouldBlock):
             # Handle graceful shutdown
             self.logger.info("Server stopped by user.")
-
-    def get_vendor_and_model(self, config: PatternExecutionConfig) -> tuple[str, str]:
-        """Get the vendor and model based on the provided configuration."""
-        vendor_name = config.vendor_name or self._default_vendor
-        if not vendor_name:
-            self.logger.debug(
-                "Vendor name is None or empty. Set to hardcoded default vendor: %s",
-                DEFAULT_VENDOR,
-            )
-            vendor_name = DEFAULT_VENDOR
-
-        model_name = config.model_name or self._default_model
-        if not model_name:
-            self.logger.debug(
-                "Model name is None or empty. Set to hardcoded default model: %s",
-                DEFAULT_MODEL,
-            )
-            model_name = DEFAULT_MODEL
-
-        return vendor_name, model_name
-
-    def _execute_fabric_pattern(
-        self,
-        pattern_name: str,
-        input_text: str,
-        config: PatternExecutionConfig | None,
-        stream: bool = False,
-    ) -> dict[Any, Any] | Generator[dict[str, Any], None, None]:
-        """
-        Execute a Fabric pattern against the API.
-
-        Separated from the tool method to reduce complexity.
-        """
-        # AC5: Client-side validation
-        if not pattern_name or not pattern_name.strip():
-            raise ValueError("pattern_name is required and cannot be empty")
-
-        # Use default config if none provided
-        if config is None:
-            config = PatternExecutionConfig()
-
-        vendor, model_name = self.get_vendor_and_model(config)
-
-        # AC3: Construct proper JSON payload for Fabric API /chat endpoint
-        prompt_data: dict[str, Any] = {
-            "userInput": input_text,
-            "patternName": pattern_name.strip(),
-            "model": model_name,
-            "vendor": vendor,
-            "contextName": "",
-            "strategyName": config.strategy_name or "",
-        }
-
-        # Add variables if provided
-        if config.variables is not None:
-            prompt_data["variables"] = config.variables
-
-        # Add attachments if provided
-        if config.attachments is not None:
-            prompt_data["attachments"] = config.attachments
-
-        request_payload = {
-            "prompts": [prompt_data],
-            "language": "en",
-            "temperature": config.temperature
-            if config.temperature is not None
-            else 0.7,
-            "topP": config.top_p if config.top_p is not None else 0.9,
-            "frequencyPenalty": config.frequency_penalty
-            if config.frequency_penalty is not None
-            else 0.0,
-            "presencePenalty": config.presence_penalty
-            if config.presence_penalty is not None
-            else 0.0,
-        }
-
-        # AC1: Use FabricApiClient to call Fabric's /chat endpoint
-        api_client = FabricApiClient()
-        try:
-            # AC4: Handle Server-Sent Events (SSE) stream response
-            response = api_client.post("/chat", json_data=request_payload)
-            response.raise_for_status()  # Raise HTTPError for bad responses
-
-            if stream:
-                # Return generator for streaming mode
-                return self._parse_sse_stream(response)
-            else:
-                # Return accumulated result for non-streaming mode
-                return self._parse_sse_response(response)
-
-        except httpx.ConnectError as e:
-            self.logger.error("Failed to connect to Fabric API: %s", e)
-            raise ConnectionError(f"Unable to connect to Fabric API: {e}") from e
-        except httpx.HTTPStatusError as e:
-            self.logger.error("Fabric API HTTP error: %s", e)
-            error_text = e.response.text
-            status_code = e.response.status_code
-            raise RuntimeError(
-                f"Fabric API returned error {status_code}: {error_text}"
-            ) from e
-        except Exception as e:
-            self.logger.error("Unexpected error calling Fabric API: %s", e)
-            raise RuntimeError(f"Unexpected error executing pattern: {e}") from e
-        finally:
-            api_client.close()
-
-    def _parse_sse_response(self, response: httpx.Response) -> dict[str, str]:
-        """
-        Parse Server-Sent Events response from Fabric API.
-
-        Returns:
-            dict[str, str]: Contains 'output_format' and 'output_text' fields.
-        """
-        # Process SSE stream to collect all content
-        output_chunks: list[str] = []
-        output_format = "text"  # default
-        has_data = False  # Track if we received any actual data
-
-        # Parse SSE response line by line
-        for line in response.iter_lines():
-            line = line.strip()
-            if not line:
-                continue
-
-            # SSE lines start with "data: "
-            if line.startswith("data: "):
-                has_data = True
-                try:
-                    data = json.loads(line[6:])  # Remove "data: " prefix
-
-                    if data.get("type") == "content":
-                        # Collect content chunks
-                        content = data.get("content", "")
-                        output_chunks.append(content)
-                        # Update format if provided
-                        output_format = data.get("format", output_format)
-
-                    elif data.get("type") == "complete":
-                        # End of stream
-                        break
-
-                    elif data.get("type") == "error":
-                        # Handle error from Fabric API
-                        error_msg = data.get("content", "Unknown Fabric API error")
-                        raise RuntimeError(f"Fabric API error: {error_msg}")
-
-                except json.JSONDecodeError as e:
-                    self.logger.warning("Failed to parse SSE JSON: %s", e)
-                    # For malformed SSE data, raise an error after logging
-                    raise RuntimeError(f"Malformed SSE data: {e}") from e
-
-        # Check if we received no data at all
-        if not has_data:
-            raise RuntimeError("Empty SSE stream - no data received")
-
-        # AC6: Return structured response
-        return {
-            "output_format": output_format,
-            "output_text": "".join(output_chunks),
-        }
-
-    def _parse_sse_stream(
-        self, response: httpx.Response
-    ) -> Generator[dict[str, Any], None, None]:
-        """
-        Parse Server-Sent Events response from Fabric API in streaming mode.
-
-        Yields chunks in real-time as they arrive from the Fabric API.
-
-        Yields:
-            dict[str, Any]: Each chunk contains 'type', 'format', and 'content' fields.
-
-        Raises:
-            RuntimeError: If the SSE data is malformed or empty.
-        """
-        has_data = False  # Track if we received any actual data
-
-        # Parse SSE response line by line
-        for line in response.iter_lines():
-            line = line.strip()
-            if not line:
-                continue
-
-            # SSE lines start with "data: "
-            if line.startswith("data: "):
-                has_data = True
-                try:
-                    data = json.loads(line[6:])  # Remove "data: " prefix
-
-                    if data.get("type") == "content":
-                        # Yield content chunks in real-time
-                        yield {
-                            "type": "content",
-                            "format": data.get("format", "text"),
-                            "content": data.get("content", ""),
-                        }
-
-                    elif data.get("type") == "complete":
-                        # Yield completion signal and end stream
-                        yield {
-                            "type": "complete",
-                            "format": data.get("format", "text"),
-                            "content": data.get("content", ""),
-                        }
-                        return
-
-                    elif data.get("type") == "error":
-                        # Yield error and end stream
-                        error_msg = data.get("content", "Unknown Fabric API error")
-                        raise RuntimeError(f"Fabric API error: {error_msg}")
-                    else:
-                        # Handle unexpected types gracefully
-                        self.logger.warning(
-                            "Unexpected SSE type: %s", data.get("type", "unknown")
-                        )
-                        raise RuntimeError(
-                            "Unexpected SSE data type "
-                            f"received: {data.get('type', 'unknown')}"
-                        )
-
-                except json.JSONDecodeError as e:
-                    self.logger.warning("Failed to parse SSE JSON: %s", e)
-                    raise RuntimeError(f"Malformed SSE data: {e}") from e
-
-        # Check if we received no data at all
-        if not has_data:
-            raise RuntimeError("Empty SSE stream - no data received")
-
-    def get_default_model_config(self) -> tuple[str | None, str | None]:
-        """Get the current default model configuration.
-
-        Returns:
-            Tuple of (default_model, default_vendor). Either or both can be None.
-
-        Note:
-            This method is primarily intended for testing and introspection.
-        """
-        return self._default_model, self._default_vendor
-
-    def _redact_sensitive_config_values(
-        self, config_data: dict[str, Any]
-    ) -> dict[str, Any]:
-        """Redact sensitive config values based on key patterns and value patterns.
-
-        Args:
-            config_data: Raw configuration data from Fabric API
-
-        Returns:
-            Configuration data with sensitive values redacted
-
-        Note:
-            - Sensitive keys with non-empty values are replaced with
-              "[REDACTED_BY_MCP_SERVER]"
-            - Values that look like API keys (based on prefixes) are also redacted
-            - Sensitive keys with empty string values are passed through as empty
-            - Non-sensitive keys are always passed through unchanged
-        """
-
-        redacted_config: dict[str, Any] = {}
-
-        for key, value in config_data.items():
-            # Check if key matches any sensitive pattern
-            key_is_sensitive = any(
-                fnmatch.fnmatch(key.upper(), pattern.upper())
-                for pattern in SENSITIVE_CONFIG_PATTERNS
-            )
-
-            # Check if value looks like an API key (only for string values)
-            value_is_api_key = False
-            if isinstance(value, str) and value:
-                value_is_api_key = any(
-                    value.startswith(prefix) for prefix in API_KEY_PREFIXES
-                )
-
-            # Redact if key is sensitive OR value looks like an API key
-            if (key_is_sensitive or value_is_api_key) and value != "":
-                # Redact non-empty sensitive values or API key values
-                redacted_config[key] = "[REDACTED_BY_MCP_SERVER]"
-            else:
-                # Pass through empty sensitive values and all non-sensitive values
-                redacted_config[key] = value
-
-        return redacted_config
-
-    def _validate_numeric_parameter(
-        self, name: str, value: float | None, min_value: float, max_value: float
-    ) -> None:
-        """Validate a single parameter against its expected range.
-
-        Args:
-            name: The name of the parameter (for error messages)
-            value: The value of the parameter to validate
-            min_value: The minimum acceptable value
-            max_value: The maximum acceptable value
-
-        Raises:
-            McpError: If the parameter is invalid
-        """
-        if value is not None:
-            try:
-                if not min_value <= value <= max_value:
-                    raise McpError(
-                        ErrorData(
-                            code=-32602,  # Invalid params
-                            message=f"{name} must be a number between"
-                            f" {min_value} and {max_value}",
-                        )
-                    )
-            except TypeError as exc:
-                raise McpError(
-                    ErrorData(
-                        code=-32602,  # Invalid params
-                        message=f"{name} must be a number between {min_value}"
-                        f" and {max_value}",
-                    )
-                ) from exc
-
-    def _validate_string_parameter(self, name: str, value: str | None) -> None:
-        """Validate a string parameter to ensure it is not empty.
-
-        Args:
-            name: The name of the parameter (for error messages)
-            value: The value of the parameter to validate
-
-        Raises:
-            McpError: If the parameter is invalid
-        """
-        if value is not None and not value.strip():
-            raise McpError(
-                ErrorData(
-                    code=-32602,  # Invalid params
-                    message=f"{name} must be a non-empty string",
-                )
-            )
-
-    def _validate_variables_parameter(self, variables: Any | None) -> None:
-        """Validate variables parameter: dict with string keys and values.
-
-        Args:
-            variables: The variables parameter to validate
-
-        Raises:
-            McpError: If the parameter is invalid
-        """
-        if variables is not None:
-            if not isinstance(variables, dict):
-                raise McpError(
-                    ErrorData(
-                        code=-32602,  # Invalid params
-                        message="variables must be a dictionary",
-                    )
-                )
-            variable_keys_and_values: list[Any] = list(
-                cast(list[Any], variables.values())
-            ) + list(cast(list[Any], variables.keys()))
-            if any(not isinstance(item, str) for item in variable_keys_and_values):
-                raise McpError(
-                    ErrorData(
-                        code=-32602,  # Invalid params
-                        message="variables must be a dictionary "
-                        "with string keys and values",
-                    )
-                )
-
-    def _validate_attachments_parameter(self, attachments: Any | None) -> None:
-        """Validate an attachments parameter to ensure it is a list of strings.
-
-        Args:
-            attachments: The attachments parameter to validate
-
-        Raises:
-            McpError: If the parameter is invalid
-        """
-        if attachments is not None:
-            if not isinstance(attachments, list):
-                raise McpError(
-                    ErrorData(
-                        code=-32602,  # Invalid params
-                        message="attachments must be a list",
-                    )
-                )
-            if any(not isinstance(item, str) for item in cast(list[Any], attachments)):
-                raise McpError(
-                    ErrorData(
-                        code=-32602,  # Invalid params
-                        message="attachments must be a list of strings",
-                    )
-                )
-
-    def _validate_execution_parameters(
-        self,
-        model_name: str | None = None,
-        _vendor_name: str | None = None,
-        temperature: float | None = None,
-        top_p: float | None = None,
-        presence_penalty: float | None = None,
-        frequency_penalty: float | None = None,
-        strategy_name: str | None = None,
-        variables: dict[str, str] | None = None,
-        attachments: list[str] | None = None,
-    ) -> None:
-        """Validate execution control parameters."""
-        # Validate temperature range
-        self._validate_numeric_parameter("temperature", temperature, 0.0, 2.0)
-
-        # Validate top_p range
-        self._validate_numeric_parameter("top_p", top_p, 0.0, 1.0)
-
-        # Validate presence_penalty range
-        self._validate_numeric_parameter(
-            "presence_penalty", presence_penalty, -2.0, 2.0
-        )
-
-        # Validate frequency_penalty range
-        self._validate_numeric_parameter(
-            "frequency_penalty", frequency_penalty, -2.0, 2.0
-        )
-
-        # Validate model_name format (basic validation - not empty string)
-        self._validate_string_parameter("model_name", model_name)
-
-        # Validate strategy_name format (basic validation - not empty string)
-        self._validate_string_parameter("strategy_name", strategy_name)
-
-        # Validate variables parameter format
-        self._validate_variables_parameter(variables)
-
-        # Validate attachments parameter format
-        self._validate_attachments_parameter(attachments)
-
-    def _merge_execution_config(
-        self,
-        config: PatternExecutionConfig | None,
-        model_name: str | None = None,
-        vendor_name: str | None = None,
-        temperature: float | None = None,
-        top_p: float | None = None,
-        presence_penalty: float | None = None,
-        frequency_penalty: float | None = None,
-        strategy_name: str | None = None,
-        variables: dict[str, str] | None = None,
-        attachments: list[str] | None = None,
-    ) -> PatternExecutionConfig:
-        """Merge execution parameters with existing config.
-
-        Parameters provided directly to the tool take precedence over
-        those in the config object.
-
-        Args:
-            config: Existing configuration (optional)
-            model_name: Model name override (optional)
-            temperature: Temperature override (optional)
-            top_p: Top-p override (optional)
-            presence_penalty: Presence penalty override (optional)
-            frequency_penalty: Frequency penalty override (optional)
-            strategy_name: Strategy name override (optional)
-            variables: Variables override (optional)
-            attachments: Attachments override (optional)
-
-        Returns:
-            Merged PatternExecutionConfig with parameter precedence
-        """
-        # Start with existing config or create new one
-        if config is None:
-            config = PatternExecutionConfig()
-
-        # Create new config with parameter precedence
-        return PatternExecutionConfig(
-            # Use the provided model_name if available; otherwise, fall back
-            # to the existing config's model_name
-            model_name=model_name or config.model_name,
-            # Use the provided vendor_name if available; otherwise, fall back
-            # to the existing config's vendor_name
-            vendor_name=vendor_name or config.vendor_name,
-            # Use the provided strategy_name if available; otherwise, fall back
-            # to the existing config's strategy_name
-            strategy_name=strategy_name or config.strategy_name,
-            # Use the provided variables if available; otherwise, fall back
-            # to the existing config's variables
-            variables=variables if variables is not None else config.variables,
-            # Use the provided attachments if available; otherwise, fall back
-            # to the existing config's attachments
-            attachments=attachments if attachments is not None else config.attachments,
-            # Use the provided temperature if not None; otherwise, fall back
-            # to the existing config's temperature
-            temperature=temperature if temperature is not None else config.temperature,
-            # Use the provided top_p if not None; otherwise, fall back
-            # to the existing config's top_p
-            top_p=top_p if top_p is not None else config.top_p,
-            # Use the provided presence_penalty if not None; otherwise, fall back
-            # to the existing config's presence_penalty
-            presence_penalty=(
-                presence_penalty
-                if presence_penalty is not None
-                else config.presence_penalty
-            ),
-            # Use the provided frequency_penalty if not None; otherwise,
-            # fall back to the existing config's frequency_penalty
-            frequency_penalty=(
-                frequency_penalty
-                if frequency_penalty is not None
-                else config.frequency_penalty
-            ),
-        )
